+++ src/Controller/AppController.php
@@ -12,30 +12,26 @@
  */
 namespace MeCms\Controller;
 
-use App\Controller\AppController as BaseController;
+use App\Controller\AppController as BaseAppController;
 use Cake\Event\Event;
 use Cake\I18n\I18n;
 
 /**
  * Application controller class
  */
-class AppController extends BaseController
+class AppController extends BaseAppController
 {
     /**
-     * Called before the controller action.
-     * You can use this method to perform logic that needs to happen before
-     *  each controller action.
+     * Called before the controller action
      * @param \Cake\Event\Event $event An Event instance
-     * @return \Cake\Network\Response|null|void
+     * @return \Cake\Network\Response|null
      * @see http://api.cakephp.org/3.4/class-Cake.Controller.Controller.html#_beforeFilter
-     * @uses App\Controller\AppController::beforeFilter()
-     * @uses isOffline()
      * @uses isSpammer()
      */
     public function beforeFilter(Event $event)
     {
         //Checks if the site is offline
-        if ($this->isOffline()) {
+        if ($this->request->isOffline()) {
             return $this->redirect(['_name' => 'offline']);
         }
 
@@ -71,18 +67,15 @@ class AppController extends BaseController
 
     /**
      * Called after the controller action is run, but before the view is
-     *  rendered.
-     * You can use this method to perform logic or set view variables that are
-     *  required on every request.
+     *  rendered
      * @param \Cake\Event\Event $event An Event instance
      * @return void
      * @see http://api.cakephp.org/3.4/class-Cake.Controller.Controller.html#_beforeRender
-     * @uses App\Controller\AppController::beforeRender()
      */
     public function beforeRender(Event $event)
     {
         //Loads the `Auth` helper.
-        //The `helper is loaded here (instead of the view) to pass user data
+        //This helper is loaded here (instead of the view) so we can pass user data
         $this->viewBuilder()->setHelpers(['MeCms.Auth' => $this->Auth->user()]);
 
         parent::beforeRender($event);
@@ -91,7 +84,6 @@ class AppController extends BaseController
     /**
      * Initialization hook method
      * @return void
-     * @uses App\Controller\AppController::initialize()
      */
     public function initialize()
     {
@@ -119,28 +111,13 @@ class AppController extends BaseController
      */
     public function isAuthorized($user = null)
     {
-        //Any registered user can access public functions
-        if (!$this->request->getParam('prefix')) {
-            return true;
-        }
-
-        //Only admin and managers can access all admin actions
+        //Only admin and managers can access admin actions
         if ($this->request->isAdmin()) {
             return $this->Auth->isGroup(['admin', 'manager']);
         }
 
-        //Default deny
-        return false;
-    }
-
-    /**
-     * Checks if the site is offline
-     * @return bool
-     * @since 2.15.2
-     */
-    protected function isOffline()
-    {
-        return $this->request->isOffline();
+        //Any registered user can access actions without prefix. Default deny
+        return !$this->request->getParam('prefix');
     }
 
     /**